возможность поставить/изменить/удалить оценку товару

+++ backend/database/migrations/2023_10_03_14_create_ratings_table.php
@@ -12,7 +12,7 @@ return new class extends Migration {
     {
         Schema::create('ratings', function (Blueprint $table) {
             $table->id();
-            $table->unsignedInteger('value');
+            $table->unsignedTinyInteger('value');
             $table->foreignId('product_id')
                 ->constrained(table: 'products')
                 ->onDelete('cascade');
